pyproject.toml
[project]
name = "diyivi"
version = "0.1.0"
description = "Backend for DIYivi"
authors = [
    {name = "Dirk Doesburg", email = "dirk.doesburg@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12,<3.13"

dependencies = [
    "aiosmtplib>=3.0.2",
    "email-validator>=2.2.0",
    "fastapi[standard]>=0.115.2",
    "pydantic-settings>=2.6.0",
    "pyjwt[crypto]>=2.9.0",
    "redis[hiredis]>=5.1.1",
    "uvicorn>=0.32.0",
]

[tool.uv]
dev-dependencies = [
    "coverage>=7.6.3",
    "fakeredis>=2.25.1",
    "mypy>=1.12.0",
    "pre-commit>=4.0.1",
    "pytest-asyncio>=0.24.0",
    "pytest>=8.3.3",
    "ruff>=0.7.0",
]
[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "N",
    "UP",
    "PIE",
    "Q",
    "PLE",
    "PLW",
    "RUF",
    "T20",
    "PT",
    "D200",
    "D201",
    "D202",
    "D204",
    "D205",
    "D206",
    "D207",
    "D208",
    "D209",
    "D210",
    "D211",
    "D212",
    "D300",
    "D301",
    "D400",
    "D401",
    "D402",
    "D403",
    "D412",
    "D414",
    "D418",
    "D419",
]

[tool.ruff.lint.isort]
known-first-party = ["app"]
